refactor: abstract functions from main process
