+++ src/main/index.ts
@@ -1,276 +1,27 @@
-import { electronApp, is, optimizer } from '@electron-toolkit/utils'
+import { electronApp, optimizer } from '@electron-toolkit/utils'
 import { LOCAL_FILE_DIRECTORY } from '@shared/constants/systems'
-import {
-  app,
-  BrowserWindow,
-  dialog,
-  ipcMain,
-  Menu,
-  MenuItemConstructorOptions,
-  shell,
-  WebContentsView
-} from 'electron'
-import { autoUpdater, UpdateDownloadedEvent } from 'electron-updater'
-import { globalShortcut } from 'electron/main'
-import { join } from 'path'
-import icon from '../../resources/icon.png?asset'
+import { app, BrowserWindow, globalShortcut } from 'electron'
+import { setupAutoUpdater } from './lib/auto-updater'
 import { runMigrate } from './lib/db/migrate'
-import {
-  copyFiles,
-  createDirectory,
-  getDirectoryTree,
-  getStat,
-  getUserDataPath,
-  openFileManagerApp,
-  renameFile
-} from './lib/ipc/file-system'
-import { getSystemInfo } from './lib/ipc/system-info'
+import { setupIPC } from './lib/ipc'
+import { createDirectory } from './lib/ipc/file-system'
+import { setupMenu } from './lib/menu'
 import { connectHttpServer } from './lib/server/app'
+import { createWindow } from './lib/window'
 
-const isMac = process.platform === 'darwin'
-
-let mainWindow: BrowserWindow | null = null
-let searchView: WebContentsView | null = null
-
-function registerSearchMenu(mainWindow: BrowserWindow) {
-  if (searchView) return
-
-  searchView = new WebContentsView({
-    webPreferences: {
-      nodeIntegration: true,
-      contextIsolation: false,
-      transparent: true,
-      preload: join(__dirname, '../preload/index.js')
-    }
-  })
-
-  searchView.setBounds({
-    x: (mainWindow?.getBounds().width ?? 0) - 418,
-    y: 0,
-    width: 418,
-    height: 86
-  })
-  mainWindow?.contentView.addChildView(searchView)
-
-  if (is.dev && process.env['ELECTRON_RENDERER_URL']) {
-    searchView.webContents.loadURL(
-      process.env['ELECTRON_RENDERER_URL'] + '/searchbar/index.html'
-    )
-    // searchView.webContents.openDevTools({ mode: 'undocked' })
-  } else {
-    searchView.webContents.loadFile(
-      join(__dirname, '../renderer/searchbar/index.html')
-    )
-  }
-
-  mainWindow?.on('resize', () => {
-    if (!mainWindow || !searchView) {
-      return
-    }
-    const bounds = mainWindow.getBounds()
-    searchView.setBounds({
-      x: bounds.width - 418,
-      y: 0,
-      width: 418,
-      height: 86
-    })
-  })
-
-  mainWindow?.webContents.on('found-in-page', (_event, result) => {
-    searchView?.webContents.send('find-in-page-result', result)
-  })
-
-  searchView.webContents.once('did-finish-load', () => {
-    searchView?.webContents.focus()
-  })
-}
-
-const menuTemplate: MenuItemConstructorOptions[] = [
-  // @ts-expect-error The code from the official documentation
-  // { role: 'appMenu' }
-  ...(isMac
-    ? [
-        {
-          label: app.name,
-          submenu: [
-            { role: 'about' },
-            { type: 'separator' },
-            { role: 'services' },
-            { type: 'separator' },
-            { role: 'hide' },
-            { role: 'hideOthers' },
-            { role: 'unhide' },
-            { type: 'separator' },
-            { role: 'quit' }
-          ]
-        }
-      ]
-    : []),
-  // { role: 'fileMenu' }
-  {
-    label: 'File',
-    // @ts-expect-error The code from the official documentation
-    submenu: [isMac ? { role: 'close' } : { role: 'quit' }]
-  },
-  // { role: 'editMenu' }
-  {
-    label: 'Edit',
-    submenu: [
-      {
-        label: 'Find',
-        accelerator: 'CmdOrCtrl+F',
-        click: () => {
-          const mainWindow = BrowserWindow.getFocusedWindow()
-          if (!mainWindow) return
-          registerSearchMenu(mainWindow)
-        }
-      },
-      { role: 'undo' },
-      { role: 'redo' },
-      { type: 'separator' },
-      { role: 'cut' },
-      { role: 'copy' },
-      { role: 'paste' },
-      // @ts-expect-error The code from the official documentation
-      ...(isMac
-        ? [
-            { role: 'pasteAndMatchStyle' },
-            { role: 'delete' },
-            { role: 'selectAll' },
-            { type: 'separator' },
-            {
-              label: 'Speech',
-              submenu: [{ role: 'startSpeaking' }, { role: 'stopSpeaking' }]
-            }
-          ]
-        : [{ role: 'delete' }, { type: 'separator' }, { role: 'selectAll' }])
-    ]
-  },
-  // { role: 'viewMenu' }
-  {
-    label: 'View',
-    // @ts-expect-error The code from the official documentation
-    submenu: [
-      { role: 'reload' },
-      { role: 'forceReload' },
-      { role: 'toggleDevTools' },
-      { type: 'separator' },
-      { role: 'resetZoom' },
-      { role: 'zoomIn' },
-      { role: 'zoomOut' },
-      { type: 'separator' },
-      { role: 'togglefullscreen' }
-    ]
-  },
-  // { role: 'windowMenu' }
-  {
-    label: 'Window',
-    submenu: [
-      { role: 'minimize' },
-      { role: 'zoom' },
-      // @ts-expect-error The code from the official documentation
-      ...(isMac
-        ? [
-            { type: 'separator' },
-            { role: 'front' },
-            { type: 'separator' },
-            { role: 'window' }
-          ]
-        : [{ role: 'close' }])
-    ]
-  },
-  {
-    role: 'help',
-    // @ts-expect-error The code from the official documentation
-    submenu: [
-      {
-        label: 'Learn More',
-        click: async () => {
-          await shell.openExternal('https://exodus.yancey.app')
-        }
-      }
-    ]
-  }
-]
-
-function setupAutoUpdater() {
-  autoUpdater.autoDownload = true
-  autoUpdater.autoInstallOnAppQuit = true
-
-  autoUpdater.on('update-available', () => {
-    console.log('🟡 Update available.')
-  })
-
-  autoUpdater.on('update-downloaded', (data: UpdateDownloadedEvent) => {
-    const result = dialog.showMessageBoxSync({
-      type: 'info',
-      title: 'Update Available',
-      message: `The new version(v${data.version}) has been downloaded. Would you like to restart and install it now?`,
-      buttons: ['Install Now', 'Later']
-    })
-
-    if (result === 0) {
-      autoUpdater.quitAndInstall()
-    }
-  })
-
-  autoUpdater.on('error', (err) => {
-    console.error('❌ Auto-updater error:', err)
-  })
-
-  autoUpdater.checkForUpdates()
-}
-
-function createWindow(): void {
-  // Create the browser window.
-  mainWindow = new BrowserWindow({
-    width: 1280,
-    height: 960,
-    show: false,
-    autoHideMenuBar: true,
-    ...(process.platform === 'linux' ? { icon } : {}),
-    webPreferences: {
-      preload: join(__dirname, '../preload/index.js'),
-      sandbox: false,
-      contextIsolation: true,
-      nodeIntegration: true
-    }
-  })
-
-  mainWindow.on('ready-to-show', () => {
-    mainWindow?.show()
-  })
-
-  mainWindow.webContents.setWindowOpenHandler((details) => {
-    shell.openExternal(details.url)
-    return { action: 'deny' }
-  })
-
-  // HMR for renderer base on electron-vite cli.
-  // Load the remote URL for development or the local html file for production.
-  if (is.dev && process.env['ELECTRON_RENDERER_URL']) {
-    mainWindow.loadURL(process.env['ELECTRON_RENDERER_URL'])
-    mainWindow.webContents.openDevTools({ mode: 'undocked' })
-  } else {
-    mainWindow.loadFile(join(__dirname, '../renderer/index.html'))
-  }
-}
-
-// This method will be called when Electron has finished
-// initialization and is ready to create browser windows.
-// Some APIs can only be used after this event occurs.
 app.whenReady().then(async () => {
   // Migrate PGlite
   await runMigrate()
 
   // Start Hono server
-  let server = await connectHttpServer()
+  const server = await connectHttpServer()
   server.start()
 
   // Create `LOCAL_FILE_DIRECTORY` directory if not exist
   await createDirectory(app.getPath('userData') + `/${LOCAL_FILE_DIRECTORY}`)
 
-  Menu.setApplicationMenu(Menu.buildFromTemplate(menuTemplate))
+  // Setup menu
+  setupMenu()
 
   // Set app user model id for windows
   electronApp.setAppUserModelId('app.yancey.exodus')
@@ -283,75 +34,7 @@ app.whenReady().then(async () => {
   })
 
   // Register IPCs
-  ipcMain.on('ping', () => console.log('pong'))
-  ipcMain.on('open-file-manager-app', async (_, path: string) =>
-    openFileManagerApp(path)
-  )
-  ipcMain.handle('get-directory-tree', async (_, path: string) =>
-    getDirectoryTree(path)
-  )
-  ipcMain.handle('get-user-data-path', async () => getUserDataPath())
-  ipcMain.handle('get-stat', async (_, path: string) => getStat(path))
-  ipcMain.handle('create-directory', async (_, path: string) =>
-    createDirectory(path)
-  )
-  ipcMain.handle(
-    'rename-file',
-    async (_, source: string, destination: string) =>
-      renameFile(source, destination)
-  )
-  ipcMain.handle('get-system-info', getSystemInfo)
-  ipcMain.handle(
-    'copy-files',
-    async (
-      _,
-      sourceFiles: {
-        name: string
-        buffer: ArrayBuffer
-      }[],
-      destinationDir: string
-    ) => copyFiles(sourceFiles, destinationDir)
-  )
-  ipcMain.handle('restart-server', async () => {
-    server.close(async () => {
-      // The optional `callback` will be called once the `'close'` event occurs.
-      // So, it's time to open a new instance.
-      server = await connectHttpServer()
-      server.start()
-      mainWindow?.webContents.send('succeed-to-restart-server')
-    })
-  })
-  ipcMain.handle('find-in-page', (_, keyword) => {
-    if (keyword === '') {
-      mainWindow?.webContents.stopFindInPage('clearSelection')
-    } else {
-      mainWindow?.webContents.findInPage(keyword)
-    }
-  })
-  ipcMain.handle('find-next', (_, keyword) => {
-    if (keyword === '') {
-      mainWindow?.webContents.stopFindInPage('clearSelection')
-    } else {
-      mainWindow?.webContents.findInPage(keyword, { findNext: true })
-    }
-  })
-  ipcMain.handle('find-previous', (_, keyword) => {
-    if (keyword === '') {
-      mainWindow?.webContents.stopFindInPage('clearSelection')
-    } else {
-      mainWindow?.webContents.findInPage(keyword, {
-        findNext: false,
-        forward: false
-      })
-    }
-  })
-  ipcMain.handle('close-search-bar', () => {
-    if (searchView) {
-      mainWindow?.contentView.removeChildView(searchView)
-      searchView = null
-      mainWindow?.webContents.stopFindInPage('clearSelection')
-    }
-  })
+  setupIPC()
 
   createWindow()
   setupAutoUpdater()
