+++ src/renderer/components/settings/settings-form/system-info.tsx
@@ -1,40 +1,41 @@
 import { FormLabel } from '@/components/ui/form'
 import { Separator } from '@/components/ui/separator'
-import { SystemInfo as SystemInfoType } from '@shared/types/system-info'
-import { Fragment, useCallback, useEffect, useState } from 'react'
-import { toast } from 'sonner'
+import { version } from '../../../../../package.json'
 
 export function SystemInfo() {
-  const [systemInfo, setSystemInfo] = useState<SystemInfoType | null>(null)
+  const { versions, platform } = window.electron.process
 
-  const getSystemInfo = useCallback(async () => {
-    if (systemInfo !== null) return
-
-    try {
-      const result: SystemInfoType =
-        await window.electron.ipcRenderer.invoke('get-system-info')
-      setSystemInfo(result)
-    } catch {
-      toast.error('Failed to retrieve system information.')
-    }
-  }, [systemInfo])
-
-  useEffect(() => {
-    getSystemInfo()
-  }, [getSystemInfo])
-
-  if (systemInfo === null) return null
   return (
     <>
-      {Object.keys(systemInfo).map((label) => (
-        <Fragment key={label}>
-          <div className="flex items-center justify-between">
-            <FormLabel>{label}</FormLabel>
-            <span className="text-ring text-sm">{systemInfo[label]}</span>
-          </div>
-          <Separator />
-        </Fragment>
-      ))}
+      <div className="flex items-center justify-between">
+        <FormLabel>Version</FormLabel>
+        <span className="text-ring text-sm">v{version}</span>
+      </div>
+      <Separator />
+      <div className="flex items-center justify-between">
+        <FormLabel>Electron</FormLabel>
+        <span className="text-ring text-sm">v{versions.electron}</span>
+      </div>
+      <Separator />
+      <div className="flex items-center justify-between">
+        <FormLabel>Chromium</FormLabel>
+        <span className="text-ring text-sm">v{versions.chrome}</span>
+      </div>
+      <Separator />
+      <div className="flex items-center justify-between">
+        <FormLabel>Node.js</FormLabel>
+        <span className="text-ring text-sm">v{versions.node}</span>
+      </div>
+      <Separator />
+      <div className="flex items-center justify-between">
+        <FormLabel>V8</FormLabel>
+        <span className="text-ring text-sm">v{versions.v8}</span>
+      </div>
+      <Separator />
+      <div className="flex items-center justify-between">
+        <FormLabel>OS</FormLabel>
+        <span className="text-ring text-sm">{platform}</span>
+      </div>
     </>
   )
 }
